04_deployment/app/requirements.txt
shiny
pandas
plotly
requests
python-dotenv
python-dateutil
sentence-transformers
sqlite-vec
